Optimize window repainting area (closes: #26).
Before, DamageReportNonEmpty was used but this was not efficient at all as it
just reports a single DamageNotify event each time the damage rectangle
changes from empty to non-empty, meaning that the damaged area will be the
full window.

Instead, use DamageReportRawRectangles to get the real damaged area of a
window. This also makes check whether a window is fully damaged more useful.