Use DamageReportDeltaRectangles to avoid being flooded with DamageNotify (refs: #26).
With DamageReportRawRectangles level, no attempt to compress out overlapping
rectangles is made, therefore many events are received and handled needlessly
(according to a test, 3 times more events are received on the same test).
Whereas with DamageReportNonEmpty level only a single event specifying the
full window region is sent thus this is not efficient for small damage
regions.