Handle _NET_WM_WINDOW_OPACITY with ARGB windows
Previously, setting an ARGB window's opacity via e.g. transset had no effect,
because this property was never evaluated. This commit fixes that.

This also fixes transparency in the git/master version of the "awesome" window
manager, because this is a reparenting WM which creates all its parent windows
with an ARGB visual (lazy programmer :-).

Signed-off-by: Uli Schlachter <psychon@znc.in>